무기 뻇어먹을 수 있도록
